window: set places sidebar to the previous location on error
Nautilus opens the location notified by GtkPlacesSidebar
without further interaction with it.

With that, when a location fails to open, Nautilus goes
back to the previous valid location but, because of we're
not interacting with the sidebar after the failure, the
sidebar was not updated to match these changes, i.e. kept
the old invalid row selected.

To fix it, set the previous valid location to the sidebar
as well if there's some error loading a location.

https://bugzilla.gnome.org/show_bug.cgi?id=734755