nautilus-view: show delete permanently on trash unsupported systems
We removed the menu item to delete permanently and instead always
allow the shortcut <shift>Delete to do so, thinking that move
to trash will be always available.

The problem is that trash is not supported on some systems, and the user
can be confused if there is no way to delete the item on the menu.

To fix that, still allow always to delete permanently with the shortcut
and show the menu item only when trash is not available.

https://bugzilla.gnome.org/show_bug.cgi?id=748692