search-provider: initialize bookmark list lazily
NautilusBookmarkList is already a singleton; since the shell provider
object is created before the application knows whether it's the primary
instance or not, loading the list in that code path makes us do ore work
than needed when we're only running as a launcher.

Just call the singleton getter every time.