window-slot: switch to list view on search
When searching, list view is a better view given that we can see
more attributes of the files, for instance, the location is
a important part for discerning files in the current folder or
in nested folders.

For that, switch to list view while searching, and disable this behavior
if the user changes the view mode while searching.