window: don't set list view as default when searching
We recently made a change that the view switchs automatically
to list view if it is searching.

But we introduced a bug, we were setting as default list view in
that case, and therefore if nautilus was closed while searching
the default view mode changed to list view.

Fix that and only set as default list view if we are not searching
and "switch to list view on search" setting is set.