view: don't crash on view destroy if no model
We stop loading the view on destroy, and we add a check on
done_loading to not do some things if we are on destroy.
But the check was checking for the slot not having a window,
which actually has if the view is being destroyed while the slot
is not, like when switching views.

To avoid this crash, mark the model as NULL before all the checking,
and check for the model, not for a window in the slot.