window-slot: don't stop the new_content load/search
We were stopping the loading and searching of the view when location
changed or when location change was canceled, etc.

Before it was fine because stopping the view was not making the
actual directory behind stop.

In previous commits we now make the model stop loading/searching
if the view stops as well, so now that this has a different meaning,
we don't want to stop the new directory, since it could be the one
that will become the current directory.

So just don't stop the current view for simple location changes.