window-slot: stop the current view loading in location change
So the current view don't keep polling files info while we try
to actually change the location.

This was causing nautilus to be slow like hell on some situations,
since the new location waits until it has it's files info ready
to actually change the view, but the old view was also polling files info.
Given that the priority of both actions are the same, even if they are
in different threads, one of those could cope the other one.

So just avoid this situation (although there is similar issues like this
with the current code...).