view: manage empty states
The NautilusView class is the abstract class from which
the different views (list & grid) inherit, in order to
display directory content.

Up to now, it was NautilusWindowSlot who was managing
the empty views, causing even more code coupling and
blurring the lines of functionality of each class.

Fix that by delegating to NautilusView the responsability
to manage the empty views, and not NautilusWindowSlot.