view: manage empty states
NautilusWindowSlot is the class responsible for managing
the empty views, causing even more code coupling and
blurring the lines of functionality of each class. Also,
it requires that the slot access the underlying model
behind NautilusView, something we're avoiding in order
to get Nautilus ready for future modifications.

Fix that by delegating to NautilusView the responsability
to manage the empty views, and not NautilusWindowSlot.