pathbar: handle context menu
NautilusView is an abstract class that manages
various context menus, depending on the view's
location, the clicked point and the implementation
details.

While this in theory provides a good isolation
from other classes, in practice NautilusView
manages the pathbar context menu, which is not
necessary, as it doesn't depend on the current
view anymore after the GAction rework.

Fix that by making NautilusPathBar manage the
context menu by itself instead of the view. To
cleanly implement that, add a new signal that
matches GtkPlacesSidebar::open-location signature,
and adapt NautilusWindow to reuse the existing
methods to handle pathbar's new signal.

https://bugzilla.gnome.org/show_bug.cgi?id=753158