window-slot: show window when view starts loading
NautilusView directly access NautilusWindow to make sure
the window is shown when the view starts loading, and not
before this.

Since we're restricting communication between NautilusView
and other widgets, the view shouldn't ask the window directly.

Fixed by making NautilusWindowSlot handle the visibility of the
window on NautilusView::begin-loading signal. Also, this patch
gets rid of an old code that updates the title of supposedly
hidden slots that were present or created before showing the
window. which is something doesn't happen anyomore.

https://bugzilla.gnome.org/show_bug.cgi?id=753356