view: handle zoom X11 events
All X11 custom keys are currently handled by NautilusWindow,
which performs a search on it's own action group as well as
the view's one to perform the corresponding action.

Since we're isolating NautilusView from other classes, the
search on the view's action group is undesired.

Fix that by making the view handle zoom-in and zoom-out actions,
instead of the window itself.

https://bugzilla.gnome.org/show_bug.cgi?id=753695