toolbar: dont update operations when no necesary
We were looping with a timeout to make sure we don't miss
any operation that it's time remaning was not well calculated.

But we were not stoping this loop when all the operations were
took into account. So remove the timeout if all the operations
are being show, cancelled, or finished.

https://bugzilla.gnome.org/show_bug.cgi?id=753728