files-view: don't add sidebar shortcuts for gtkfilechooser
That was used when the action "Move to..." or "Copy to..."
since it was using gtkfilechooser to choose the destination.
We were adding a sidebar shortcut folder for every slot that was
opened, that means that if we had a few tabs, the sidebar started
to look crowed. Also the sidebar is no longer consistent, since
it's diffenrent thant the state of the sidebar in the main Nautilus
window.

Designers agree that is better to make the sidebar less crowed and
don't make it incosistent, since the value of adding a shorcut for
opened locations doesn't balance the inconsistency and crowdness of
the sidebar.