slot: use location property
We were using a custom signal that reports the previous
location and the new one, given that was needed by files-view
for its bookmarks in the file chooser.
Previous patch removed that, so now the only client is the application
which updates the dbus locations available to FileManager dbus.

We want to have some consistency between the handling of the
views/slots/windows, and we are moving forward to use properties
instead of signals since makes it clearer and cleaner, and allow
bindings and other nice things.

For that, match what the view does and expose a location property
on the slot.

In future, if the window has some client that listen to the current
viewed location, a property binding to the active slot would do.

In this case, we can't bind the property to the one of the view, since
there is still some management done on the slot, but this will change
soon when we move the remaining location management to the view itself.