files-view: don't open folder with timer for drag&drop
We were opening the hovered folder after a timeout when performing a
drag and drop operation.
However, this causes that the user could lose track of were the user was
and fail the dnd operation, or even worse, drop in a unknow location and
makes the whole operation rather anoying.
Also looks like could be a problem for disabled people.

So instead just don't open locations on hover for the main view.

Pathbar remains opening with the timer on hover, since the pathbar
itself doesn't change and therefore the user don't lose the track.

https://bugzilla.gnome.org/show_bug.cgi?id=727790