list-view: disable sorting on recent and search
Following what we did in the canvas view, and what the file
chooser is doing, we disallow the manual sorting on recent
and search for list view as well, since it kinda defeats the
point to allow a diferent sorting in those situations.

However, for search we can still want some different sorting,
but let's try this way and also be consistent with filechooser.

https://bugzilla.gnome.org/show_bug.cgi?id=751336