files-view: ensure sort criteria correcteness
Since last commit it's not allowed o change the sort criteria
on search and recent, but the user could changed it before updating
nautilus or some application could modified the metadata itself,
making the user incapable to set the correct values.
To fix that, override at runtime the sort criterion for recent and
search, but withouth modifying the metadata.

https://bugzilla.gnome.org/show_bug.cgi?id=751336