dnd: add setting for opening on hover on dnd
In previous commits we removed opening automatically a folder
when performing a dnd operation and hovering it, since it
is rather difficult to control. Increasing the timeout doesn't
fix the problem, since we can't know how much a person needs to
control its mouse or see the files in the current folder.
Basically it performs an action even if the user is still, which
is a problem.

But seems some users were used to it, so instead of removing it
altogether add a gsetting key to allow its use.
The key will be set as false by default to avoid making the experience
difficult out of the box.

https://bugzilla.gnome.org/show_bug.cgi?id=754454