slot: clean up
Clean up a bit more...given that we have to keep the code to decide
how to handle the location the application is giving to us, and
since we don't know if it is a regular file, a mount file etc
which needs management, we still need its info.
Doing it in the view will actually give a lot of burden to it,
and also is not part of the view decide how to handle kind
of files. So we have to decide doing it on the slot or in the
application.
At some point I would like to move it to the application, but
then we will need a way to make the slot wait to set a real location.