window-slot: ensure location when creating query
We were not setting a location when showing the query for first
time, that means that when query changed, it didn't have a location
and therefore the search failed, until the next iteration that the
location was set.
To fix it, set the query location when showing the query for first time.