window-slot: check wheter to close search or not in location change
We weren't closing the search at all when location changed. Previously
the work on the latest months, the search was being close if the
query was non empty and the user change location, and if the query
was empty, it was keeping the search visible.

Before it was done keeping track wheter keep search visible between
location changes, and doing a special check for that. However now that
we removed most of that code, we need a new way to deal with it withouth
relying on the private structure and be base in actions or properties
as we have been trying to do lately.

For that, improve some of the checks that were being done (the code was
sometimes randomly doing some checks which prevented it to work
differently and was creating races) and use the action to keep track
of all the searching part on the slot.