files-view: take a ref to source object when loading directory
We were calling load_directory with the current model, but the view
expect to own it's own reference to the model, and the caller to own
it's own one to the source object.
However, when we use that internally, we were just using the current
model withouth taking an aditional reference, which was causing a
use-after-free crash.

To avoid it, if we are going to us public functions internally, take a
reference to the source object like a external client would have.

https://bugzilla.gnome.org/show_bug.cgi?id=754310