canvas-container: fix bounding box on different pixels per unit
We were returning the bounding box without taking care of the pixels
per unit, so the clients of it were using the wrong bounding box
calculation. For instance the rename popover was wrong calculated
for the canvas view if the pixel per unit was different than 1.

The bounding boxes are usually not calculated with pixels per unit
adjustment, but is very convenient to have it here, so add a comment
explaining it.

https://bugzilla.gnome.org/show_bug.cgi?id=754620