file-operations: fix crash when copying
Commit 2db9a295 introduced an obvious crash, not sure how
I managed to make it in if it was part of the tests and seems
it was working for other people.
Anyway, we were using the wrong argument, which pointed
to a wrong address when reporting operations, and that was
causing a crash.

To fix it, remove the  spurious argument.

https://bugzilla.gnome.org/show_bug.cgi?id=753282