window: save selected item for sidebar properties action
If not, we don't know the item that was selected and therefore
can't show the properties window for locations different than
the current one selected in the sidebar.
Before was working because sidebar was using GtkTreeView and
it always had an item selected when right click, therefore
gtk_sidebar_get_location was actually the file of the item
the user did right click.

https://bugzilla.gnome.org/show_bug.cgi?id=754677