file, file-undo-operations: Let renames on Google Drive be undone
We really should be using standard::display-name when renaming files -
g_file_get_basename is simply not the right thing to use. Moreover,
in case of Google Drive, the URI might not change with the display
name and we can not get the old display name from the old GFile.

Therefore, let's explicitly keep track of the old and new display
names to address these issues.

https://bugzilla.gnome.org/show_bug.cgi?id=754785