desktop: avoid to raise the window on the stack
If we don't pass DONT_MAKE_ACTIVE flag to the window, it will
raise the window with a gtk_present_window () call, which will
actually hide the desktop window from the background.

This fixes the icons not appearing until the next map () of the
desktop window, that lowers again the window on the stack.

https://bugzilla.gnome.org/show_bug.cgi?id=754360