canvas-container: avoid creating layout if not grid
When relayouting we want to finish adding and layouting
the previous layout, but sometimes we can try to layout
when it's actually not possible to create a grid, for
instance if the size of the desktop is not enough or is
not yet allocated.
To prevent that, return early if we try to relayout in
that case and schedule a relayout.

https://bugzilla.gnome.org/show_bug.cgi?id=755582