window: avoid crash when no dnd items
Sometimes is reported that there are not items as sources.
That makes nautilus crash trying to iterate betweeen them
in nautilus_drag_default_drop_action_for_icon.
Even if the fix should be there, it's a situation that
shouldn't happen at all, so I leave it as a responsability
of the caller.