settings: open folders on dnd by default
I have been going back and forward on this. And even if
I strongly think this is not good enough to be in the default
installation of Nautilus, removing 'features' or behaviors that
users were used to without providing a good alternative is even worse.
Probably Nautilus didn't have such high standards when this was
implemented, and this behavior was accepted even that emails from
usability and accessibility were asking to not do it, but it was done
given that no alternative was possible because of Apple patents on
'spring folders'.

So final decision from my side and after discussing it with some
people, is to continue shipping it by default until we have a better
solution.

The setting will be still there as 'open-folder-on-dnd-hover' and
can be modified by the user using dconf.

Discussions for better solution will be followed in
https://bugzilla.gnome.org/show_bug.cgi?id=754455