directory-async: keep directory alive
Lately we did some changes on the view/slot/window/app
interaction with the directory/model ownership. That means
that now most of the times the directory/model is freed when
it should.
In this case, when loading the children of a directory reports
an error, say the user doesn't have permission to access the
directory, the view unref the directory in the signal callback,
and any further interaction with it by the directory itself after
signaling becomes accessing invalid memory.

To avoid this, the directory should have a ref on itself for any
signal that can cause the owners of it to unref.

https://bugzilla.gnome.org/show_bug.cgi?id=755207