files-view: enable properties on all cases
We changed the meaning of remote from being only network://
to be a lot of schemes.
Thing is, the properties action was disabled for "remote", and
now that means lot of locations.
I think there is not enough reason to disable properties altogether,
if the network:// file doesn't provide enough info, probably it should.