file: get filesystem type asynchronously
We were peeking the filesystem type synchronously since nautilus
was checking for some file infos that are only required once
synchronously. However the call for the filesystem is usually slow
which makes nautilus main thread hang.

To avoid I/O in the main thread, make the filesystem type request
asynchronously and part of all the others attributes for the file.

https://bugzilla.gnome.org/show_bug.cgi?id=756280