application: specify cwd
We are managing all command line options in the main instance.
That works always correctly except when resolving relative paths,
which are relative the local instance, not the main one.

To fix it, set a "cwd" option in the local instance to ensure the
relative file paths are resolved in the main instance based on the local
instance.

https://bugzilla.gnome.org/show_bug.cgi?id=756688