desktop file: specify new window action
So Gnome Shell is aware of it and can activate it.
Until now Gnome Shell was guessing it was able to open a new
window given the app.new-window action. However, what Gnome Shell
does in this case is just call activate of the application, which
is not the "new-window" action for nautilus, and instead only presents
the current window.

One can argue that Gnome Shell should not try to guess the available
actions to, after that, instead of using them, just activate the
application.

https://bugzilla.gnome.org/show_bug.cgi?id=756370