desktop: undecorate the window before positioning
Starting with gtk-+3.19, gtk+ tries to compensate for the client-side
decorations when moving/resizing top level windows.

With this, the Nautilus desktop window is misplaced because at the time
it's positioned, gtk+ cannot determine it will be undecorated
eventually, as both the gtk_window_set_decorated() and the type hint
(_NET_WM_WINDOW_TYPE) are set after gtk_window_move().

To avoid this, invoke the window positioning after
gtk_window_set_decorated() so that gtk+ is aware that the window is not
decorated and doesn't apply the offset to compensate for client-side
decorations.

Bugzilla: https://bugzilla.gnome.org/show_bug.cgi?id=757471