Added custom shortcuts support to launch scripts
The problem is that Nautilus doesn't permit to map shortcuts to launch
the scripts in the scripts folder.

I added a function to read from a configuration file the shortcuts to
map the scripts.
When the scripts are added to the context menu they will have associated
the shortcuts written in the config file.

https://bugzilla.gnome.org/show_bug.cgi?id=757766