toolbar: avoid empty operations popover
We need to check that all windows have the popover hidden
to remove the operations, if not, what happens is that if
one of the windows has it hidden it removes its operations
and all the other popovers become empty.

The code is not very beautiful since we have to access all
the toolbars of all the windows, which is kind of breaking
the design, but well... creating a "operations toolbar manager"
just for this looks overkill.

https://bugzilla.gnome.org/show_bug.cgi?id=756560