window: explicitily set active window
We were relying on getting the current active window
from gtk to switch open locations, but what people found
is that some times another window change its location instead
of the current one.
I couldn't figure out the root of the problem, so I'm going
to pass explicitly the window on the callers as previous 3.18
we were doing.
Would be cool to find the root of the problem though...

https://bugzilla.gnome.org/show_bug.cgi?id=756499