window-slot: don't ref all views on creation
This was leaking and making that no view was finalized, leading
to all kind of problems and crashes due to signals and files updating
the dangling view, and not letting files to be released.
This was creating multiple non-related reports on Fedora Tracker making
nautilus one of the top crashers.

Downstream tracker bug:
https://bugzilla.redhat.com/show_bug.cgi?id=1277625

The upcoming patch is also necessary for a complete fix.