file: fix wrong date calculation
We were checking how many days ago the file was modified by
checking directly the current time with the file modification
time, which is wrong if the file was modified a few hours ago
but the day already passed.
What happened is that it was showing the time as in if it was
modified in the same day, instead of saying "Yesterday".

To fix it, check the time difference with the midnight time.

https://bugzilla.gnome.org/show_bug.cgi?id=757272