files-view: unref action group
We were leaking it, causing that the view was not free if
there were some templates, since templates menu items were
holding a ref to the view and are only freed when the closure
of the action gets freed, which is the view action group.
This was making the view to get never finalized, and therefore
making all kind of problems afterwards.

Upcoming patch will also make template not hold a ref to the view,
even if the issue is fixed now, I think templates menu items holding
a reference to the view is wrong.