canvas-container: consider margins when calculating icon bounding box
Shells can fix docks, panels and other similar things which reduce the
available space to draw on. This is called the canvas margin.

When positioning things relative to icons within containers we need to
adjust by this margin, otherwise they will be shifted away from the
target. Do this when we calculate the bounding box.

https://bugzilla.gnome.org/show_bug.cgi?id=759002