mime-actions: use explicit slot for opening locations
Even if we wanted to always use the active window as the default
action to open locations, seems gtk_get_active_windows doesn't work
as expected for us. Until we figure out the problem, pass explicitly
the target slot that the caller provides so users don't experience
the location opening in a different window than the expected.

https://bugzilla.gnome.org/show_bug.cgi?id=756499