docs: add an OVERVIEW file
It has been difficult to know the ownership and other things of
nautilus over the time. To not spend again so much time on it, it would
be helpful to have a short overview of nautilus internals.

For that, add a markdown file that explains general concepts and
behaviors of the internals of nautilus.

For now I explained the file, directory and view ownership, to avoid
future confusions and leaks which are kind of catastrophic in Nautilus.